refactor: moved button to buttons folder
