+++ src/components/Buttons/Button/index.tsx
@@ -1,4 +1,3 @@
-import { AiOutlinePlus } from 'react-icons/ai'
 import styles from './styles.module.scss'
 
 interface IButton {
@@ -9,10 +8,6 @@ interface IButton {
 const Button = (props: IButton) => {
   return (
     <button className={styles.button} onClick={props.onClick}>
-      <div className={styles.icon}>
-        <AiOutlinePlus/>
-      </div>
-
       <p className={styles.button_text}>{props.text}</p>
     </button>
   )
