+++ src/components/Buttons/Button/styles.module.scss
@@ -1,0 +1,29 @@
+@import '../../../styles/break-points.scss';
+
+.button{
+    position: relative;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding: 0px 32px;
+  height: 45px;
+  width: fit-content;
+  background-color: #fff;
+  border: 2px solid #65dcc7;
+  border-radius: 8px;
+  box-shadow: 1px 2px 5px 0px rgba(0, 0, 0, 0.2);
+  transition: all 200ms ease-in-out;
+  cursor: pointer;
+  &:hover {
+    background-color: #65dcc7;
+    .button_text {
+      color: #fff;
+    }
+  }
+  .button_text {
+    color: #587169;
+    font-weight: bold;
+    font-size: 16px;
+    text-transform: uppercase;
+  }
+}
