+++ src/components/Buttons/index.ts
@@ -1,1 +1,2 @@
 export * from './Submit'
+export * from './Button'
