+++ src/components/index.tsx
@@ -1,2 +1,2 @@
-import Button from "./Button";
+import Button from "./Buttons/Button";
 export { Button };
